backend/backend/requirements.txt
asgiref==3.9.1
django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
pyjwt==2.9.0
sqlparse==0.5.3

Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==25.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
requests==2.32.3
